test(parser): change tests in parser to scope analysis
