+++ packages/parser/tests/lib/comments.ts
@@ -1,6 +1,9 @@
 import fs from 'fs';
 import glob from 'glob';
-import * as testUtils from '../../tools/test-utils';
+import {
+  createSnapshotTestBlock,
+  formatSnapshotName
+} from '../tools/test-utils';
 import { ParserOptions } from '../../src/parser-options';
 
 const FIXTURES_DIR =
@@ -21,8 +24,8 @@ describe('Comments', () => {
       }
     };
     it(
-      testUtils.formatSnapshotName(filename, FIXTURES_DIR),
-      testUtils.createSnapshotTestBlock(code, config)
+      formatSnapshotName(filename, FIXTURES_DIR),
+      createSnapshotTestBlock(code, config)
     );
   });
 });
